requirements.txt
discord.py == 2.3.2
PyNaCl == 1.5.0
customtkinter == 5.2.1
packaging == 23.2